### 2018-10-03 aaa9747
making slight updates to bib
