### 2018-10-04 45483cc
fixed extra curly bracket in bib file
