### 2018-10-04 #993 @wino6687
I had an extra '{ ' in my bib file. I have pushed a fix to the repo. Is there a way to trigger @whedon to compile it again? Thanks!
