### 2018-10-04 be49649
adding a function to open 19 and 37 tb files and get swe numpy array returned
