### 2018-10-04 4312112
removing mapbox functionality, prematurely added to master;
