+++ swepy/swepy.py
@@ -11,8 +11,6 @@ from nco import Nco
 import os
 from tqdm import tqdm
 import glob
-from mapboxgl.utils import *
-from mapboxgl.viz import *
 
 nco = Nco()
 
@@ -444,49 +442,3 @@ class swepy():
 
         # difference for SWE
         return self.safe_subtract(tb_19H_long, tb_37H_long)
-
-    def plot_a_day(self, token):
-        '''read tb,x,y data from final files,
-        with the purpose of plotting.'''
-        fid_19H = Dataset(self.concatlist[0], "r", format="NETCDF4")
-        fid_37H = Dataset(self.concatlist[1], "r", format="NETCDF4")
-
-        x = fid_19H.variables['x'][:]
-        y = fid_19H.variables['y'][:]
-
-        tb_19H = fid_19H.variables['TB'][:]
-        tb_37H = fid_37H.variables['TB'][:]
-        tb_37H = block_reduce(tb_37H, block_size = (1,2,2), func = np.mean)
-
-        tb = self.safe_subtract(tb_19H, tb_37H)
-        lats = np.zeros((len(y), len(x)), dtype=np.float64)
-        lons = np.zeros((len(y), len(x)), dtype=np.float64)
-        grid = ease2Transform.ease2Transform(gridname=fid_19H.variables["crs"].long_name)
-        one_day = tb[0,:,:]
-        df = pd.DataFrame(columns = ['lat', 'lon', 'swe'])
-        for i, xi in enumerate(x):
-            for j, yj in enumerate(y):
-                row, col = grid.map_to_grid(xi, yj)
-                lat, lon = grid.grid_to_geographic(row, col)
-                lats[j, i] = lat
-                lons[j, i] = lon
-                #df = df.append({'lat': lats[j][i], 'lon': lons[j][i], 'swe':one_day[j-2][i-2]}, ignore_index = True)
-        for i in range(len(one_day[:,1])):
-            for j in range(len(one_day[1,:])):
-                df = df.append({'lat': lats[i][j], 'lon': lons[i][j], 'swe':one_day[i][j]}, ignore_index = True)
-        print(one_day)
-        os.chdir(self.working_dir)
-        df_to_geojson(df, filename = 'swe_1day.geojson',properties = ['swe'],lat = 'lat', lon = 'lon')
-        measure = 'swe'
-        color_breaks = [round(df[measure].quantile(q=x*0.1), 2) for x in range(1,9)]
-        color_stops = create_color_stops(color_breaks, colors='YlGnBu')
-        # Create the viz from the dataframe
-        viz = CircleViz('swe_1day.geojson',
-                        access_token=token,
-                        color_property = "swe",
-                        color_stops = color_stops,
-                        center = (self.center),
-                        zoom = 3,
-                        below_layer = 'waterway-label')
-
-        viz.show()
