### 2018-10-04 51107f2
tossed the center calculation that was preventing tests from completing when boundng coords werent passes
