### 2018-10-06 d5aedd2
adding some documentation inside swepy.py, real basic
