### 2018-10-08 f5c7cf6
catching swepy up before new release with open_swe function for new conda release
