### 2018-10-08 979c435
updated the orientation guidlines for subsetting files
