### 2018-10-08 #993 @wino6687
@Chilipp would you be willing to do this review? I noticed you have experience working with climate data, and possibly remotely sensed stuff. Maybe this would be a good fit? Thanks!
