### 2018-10-08 2f712a6
slowly fixing post 2000s 25km web scraping now that v1.3 changed file paths
