### 2018-10-08 8c2abad
adding conda_upload.sh file for testing purposes
