### 2018-10-08 e866461
changing file path to conda_upload.sh in travis.yml
