### 2018-10-09 69703e7
fixed critical issue in get_file that was breaking ability to scrape high res files
