+++ swepy/swepy.py
@@ -276,17 +276,17 @@ class swepy():
                 date2 = date - timedelta(days = 1)
             else:
                 date2 = date
+            if date in [datetime(2003,11,6), datetime(2004,4,9)]:
+                sensor = 'F14'
+            if date in [datetime(2006,11,4), datetime(2006,12,1), datetime(2008,2,26)]:
+                sensor = 'F15'
+                ssmi_s = "SSMI"
+            if date in [datetime(2005,5,12),datetime(2006,2,4),datetime(2008,1,2),datetime(2008,2,26)]:
+                pass1 = 'E'
         if self.grid == 'T':
             pass1 = 'A'
         else:
             pass1 = 'M'
-        if date in [datetime(2003,11,6), datetime(2004,4,9)]:
-            sensor = 'F14'
-        if date in [datetime(2006,11,4), datetime(2006,12,1)]:
-            sensor = 'F15'
-            ssmi_s = "SSMI"
-        elif date in [datetime(2005,5,12),datetime(2006,2,4),datetime(2008,1,2)]:
-            pass1 = 'E'
 
         file = {
             "protocol": "http" if self.local_session else "https",
