### 2018-10-09 8eb1220
updating travis yml to avoid trying to upload to conda automatically for now;
