### 2018-10-12 5c658d6
updating meta.yaml and setup.py before releasing new version;
