+++ meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: "swepy"
-  version: "1.1.5"
+  version: "1.1.6"
 
 source:
-  git_rev: v1.1.5
+  git_rev: v1.1.6
   git_url: https://github.com/wino6687/SWEpy
 
 requirements:
