### 2018-10-12 71868e0
updating meta.yaml and setup.py before releasing new version;
