+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="swepy",
-    version="1.1.6",
+    version="1.1.7",
     author="William Norris",
     author_email="wino6687@colorado.edu",
     description="A python package for obtaining and manipulating Tb files from the MEaSUREs database",
