### 2018-10-13 #993 @usethedata
@wino6687 Sorry I missed the notification 3 days ago.  Yes, I'll take a crack at this.  It will be the first one I've done, so some instructions on the mechanics will be useful.
