### 2018-10-14 #993 @wino6687
Hi @usethedata, thanks for agreeing to do this review. I believe once an editor is assigned there will be a check list here in the issue, but this is my first submission so I am not 100% on that. I do see that other recent submissions that have assigned editors and reviewers have the check list locally in the issue.
