### 2018-10-14 #993 @usethedata
I'm going off https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and https://joss.readthedocs.io/en/latest/review_criteria.html.  That will at least be a start, I'm sure folks will let me know if I miss anything.
