### 2018-10-14 #1018 @whedon
[REVIEW]: SWEpy: A Python Library for Quick and Easy Access and Manipulation of MEaSUREs CETB files for SWE analysis
**Submitting author:** @wino6687 (<a href="http://orcid.org/0000-0002-6772-9600">will norris</a>)
**Repository:** <a href="https://github.com/wino6687/SWEpy" target ="_blank">https://github.com/wino6687/SWEpy</a>
**Version:** v1.1.7
