### 2018-10-14 #1018 @arfon
@wino6687 - it looks like your paper doesn't have any references, or if they are, they're not being compiled in. You can read how to cite works in Pandoc [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
