### 2018-10-14 #1018 @usethedata
@arfon What's the correct way to document what I don't see as a COI so that it's out there?  In terms of the stated COI policy, we're fine.  I've never been a collaborator with either of the authors, nor have I ever collaborated directly with the Earth Lab at UC Boulder.   People who report to me are collaborators with the National Snow and Ice Data Center (NSIDC) at UC-B, which is the source of the data referenced by this software.  But nobody that works for me is a collaborator (or has been a collaborator) with UC-B Earth Lab in the past 4+ years.  I'm the PI for one of the NASA DAACs (ORNL DAAC), and part of NSDIC funding is a separate contract from NASA, but under the same program.
