### 2018-10-14 #1018 @arfon
> @arfon What's the correct way to document what I don't see as a COI so that it's out there?

I think you've just done it 😁. Thanks for disclosing this - and please proceed with the review.
