### 2018-10-14 #1018 @wino6687
@arfon, thanks for pointing that out! I do reference the NSIDC database that is used, I will get that fixed now.
