### 2018-10-14 090aa5b
making some changes to bibliography file to try and get it to properly compile
