### 2018-10-14 bc84634
added missing comma in bib, hoping that compiles it!
