+++ joss/bibliography.bib
@@ -2,5 +2,5 @@
       author = {Brodzik, M. J., D. G. Long, M. A. Hardman, A. Paget, and R. Armstrong},
       title = {MEaSUREs Calibrated Enhanced-Resolution Passive Microwave Daily EASE-Grid 2.0 Brightness Temperature ESDR, Version 1.3},
       doi = {10.5067},
-      howpublished = {\url{https://doi.org/10.5067/MEASURES/CRYOSPHERE/NSIDC-0630.001}}
+      howpublished = {\url{https://doi.org/10.5067/MEASURES/CRYOSPHERE/NSIDC-0630.001}},
 }
