### 2018-10-14 ebac8f1
changed bib format to be manual vs misc
