### 2018-10-14 875ec10
changed misc citation to not have a doi field. hoping this fixes it...
