### 2018-10-15 eef4e47
fixed inline citation for joss paper, should work now!!
