### 2018-10-15 #1018 @wino6687
Alright, the citation is fixed, sorry it took so many iterations!
