### 2018-10-15 bf7f4f4
fixing open_swe function to work for high and low res files
