### 2018-10-16 5bf38b9
fixed open_swe function
