### 2018-10-18 f2ec461
updated yaml to v1.1.7
