+++ meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: "swepy"
-  version: "1.1.6"
+  version: "1.1.7"
 
 source:
-  git_rev: v1.1.6
+  git_rev: v1.1.7
   git_url: https://github.com/wino6687/SWEpy
 
 requirements:
