### 2018-10-23 #1018 @arfon
> Hi @arfon, I updated the current git release over the last couple weeks to v1.1.7. Is there a way to update the JOSS review to v1.1.7?

Done.
