### 2018-11-11 04342c4
updated readme to be easier to understand in the introduction. Goal is to be easy to use, so the documentation has to be worded in an understandable manner
