+++ README.md
@@ -9,7 +9,8 @@
 
 # SWEpy Quick Start Guide
 ### For Full Documentation, Please see the [Wiki](https://github.com/wino6687/SWEpy/wiki)!
-SWEpy is a python library designed to give you quick and easy access to temperature brightness imagery stored in the MEaSUREs dataset in the NSIDC 0630 database. SWEpy contains tools to web scrape, geographically subset, and concatenate files into time cubes. There is an automated workflow to scrape long time series while periodically stopping to geographically subset and concatenate files in order to reduce disk impact.
+
+SWEpy is a Python library designed to simplify access to a [passive microwave brightness temperature dataset](https://doi.org/10.5067/MEASURES/CRYOSPHERE/NSIDC-0630.001) available at the National Snow and Ice Data Center (NSIDC). This dataset contains Northern and Southern hemisphere imagery along with Equatorial imagery, and can be quite useful in analyzing snow water equivalent (SWE) over large spatial extents. SWEpy contains tools to web scrape, geographically subset, and concatenate files into time cubes. There is an automated workflow to scrape long time series while periodically stopping to geographically subset and concatenate files in order to reduce disk impact.
 
 ## Setup:
 
@@ -166,4 +167,4 @@ Brodzik, M. J., D. G. Long, M. A. Hardman, A. Paget, and R. Armstrong. 2016. MEa
 
 
 ### Acknowledgements:
-I want to thank Davey Lovin and David Nyberg for their help in conceiving the original project and developing the original prototype! 
+I want to thank Davey Lovin and David Nyberg for their help in conceiving the original project and developing the original prototype!
