### 2018-11-12 64446a2
updating yaml, cleaning up, adding code-cov
