### 2018-11-12 2a988bd
changing travis to build from yml file instead
