### 2018-11-12 3a976f4
updating travis to include codecov
