### 2018-11-12 5199153
removing verbose command on pytest call
