### 2018-11-12 8cd7084
removing "=" from travis yaml
