### 2018-11-12 b38b989
installing pytest-cov to see if this fixes erros
