+++ .travis.yml
@@ -28,7 +28,7 @@ before_script:
   # run mock server in order to test scraping
   - python swepy/tests/mock_server.py &
 script:
-  - pytest --cov swepy
+  - pytest --cov=swepy
 
 after_success:
   - codecov
