### 2018-11-27 41269d2
added some comments, cleaning up master before branching to dev
