### 2018-12-02 7c25ff5
added DOI to readme
