### 2018-12-02 fdff078
added DOI to readme
