### 2018-12-06 #1018 @arfon
> Hi @kyleniemeyer and @arfon, is there anything else I should be doing right now to help with the process? Thanks!

Thanks for the bump. This got lost in the handoff between editors.
